Auction i CategoryRepository

+++ src/main/java/com/example/auction/repository/AuctionRepository.java
@@ -3,5 +3,8 @@ package com.example.auction.repository;
 import com.example.auction.model.AuctionModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuctionRepository extends JpaRepository<AuctionModel, Integer> {
+import java.math.BigDecimal;
+
+public interface AuctionRepository extends JpaRepository<AuctionModel, Long> {
+
 }
